Move hooks to own folder outside components
