+++ frontend/components/Dashboard/Dashboard.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react'
 import { useRouter } from 'next/router'
 import Nav, { navConstants } from './Nav'
 import Header from './Header'
-import useWindowSize from '../Hooks/useWindowSize'
+import useWindowSize from '../../hooks/useWindowSize'
 
 interface Props {
   children: React.ReactNode
