+++ frontend/components/Modal/Modal.tsx
@@ -4,7 +4,7 @@ import ModalHeader from './ModalHeader'
 import ModalBody from './ModalBody'
 import ModalFooter from './ModalFooter'
 import modalConstants from './modalConstants'
-import useFocusTrap from '../Hooks/useFocusTrap'
+import useFocusTrap from '../../hooks//useFocusTrap'
 import { white } from '../../utils'
 
 interface Props {
